don't require pandoc to install pypandoc
This changes closes #25 by allowing for pypandoc to install
without pandoc being present on the system.  This allows for
the library to be installed but for clients to handle the
fact that pandoc is not installed in their applications at
runtime.

If pandoc is not present during installation, the warning
from pypandoc is still printed informing users how pandoc
can be installed.